remove support for publishing with the old Android Gradle APIs
