--- plugin/src/integrationTest/fixtures/passing_android_7_1_1_project/build.gradle
@@ -1,22 +1,0 @@
-buildscript {
-    repositories {
-        mavenLocal()
-        mavenCentral()
-        google()
-    }
-
-    dependencies {
-        classpath "com.vanniktech:gradle-maven-publish-plugin:${System.getProperty("com.vanniktech.publish.version")}"
-        classpath "com.android.tools.build:gradle:7.1.2"
-        classpath "org.jetbrains.dokka:dokka-gradle-plugin:1.6.10"
-    }
-}
-
-apply plugin: "com.android.library"
-apply plugin: "com.vanniktech.maven.publish"
-
-android {
-    compileSdkVersion 29
-}
-
-apply from: "maven-publish.gradle"
