--- plugin/src/integrationTest/fixtures/passing_android_7_1_1_project/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.vanniktech.maven.publish.test" />
