--- plugin/src/integrationTest/fixtures/passing_android_7_1_1_release_variant_project/expected/test-artifact-1.0.0.pom
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd" xmlns="http://maven.apache.org/POM/4.0.0"
-    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <!-- This module was also published with a richer model, Gradle metadata,  -->
-  <!-- which should be used instead. Do not delete the following line which  -->
-  <!-- is to indicate to Gradle or any Gradle module metadata file consumer  -->
-  <!-- that they should prefer consuming it instead. -->
-  <!-- do_not_remove: published-with-gradle-metadata -->
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.example</groupId>
-  <artifactId>test-artifact</artifactId>
-  <version>1.0.0</version>
-  <packaging>aar</packaging>
-  <name>Gradle Maven Publish Plugin Test Artifact</name>
-  <description>Testing the Gradle Maven Publish Plugin</description>
-  <url>https://github.com/vanniktech/gradle-maven-publish-plugin/</url>
-  <inceptionYear>2018</inceptionYear>
-  <licenses>
-    <license>
-      <name>The Apache Software License, Version 2.0</name>
-      <url>https://www.apache.org/licenses/LICENSE-2.0.txt</url>
-      <distribution>repo</distribution>
-    </license>
-  </licenses>
-  <developers>
-    <developer>
-      <id>vanniktech</id>
-      <name>Niklas Baudy</name>
-      <url>https://github.com/vanniktech/</url>
-    </developer>
-  </developers>
-  <scm>
-    <connection>scm:git:git://github.com/vanniktech/gradle-maven-publish-plugin.git</connection>
-    <developerConnection>scm:git:ssh://git@github.com/vanniktech/gradle-maven-publish-plugin.git</developerConnection>
-    <url>https://github.com/vanniktech/gradle-maven-publish-plugin/</url>
-  </scm>
-</project>
