--- plugin/src/integrationTest/fixtures/passing_android_7_1_1_release_variant_project/src/main/java/com/vanniktech/maven/publish/test/TestActivity.java
@@ -1,18 +1,0 @@
-package com.vanniktech.maven.publish.test;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-/**
- * Just a test Activity with Javadoc.
- */
-public class TestActivity extends Activity {
-  /**
-   * Main method which does something.
-   *
-   * @param args Command-line arguments passed to the program.
-   */
-  public void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-  }
-}
