+++ plugin/src/integrationTest/fixtures/passing_android_project/build.gradle
@@ -6,8 +6,9 @@ buildscript {
     }
 
     dependencies {
-       classpath "com.vanniktech:gradle-maven-publish-plugin:${System.getProperty("com.vanniktech.publish.version")}"
-       classpath "com.android.tools.build:gradle:7.0.0"
+        classpath "com.vanniktech:gradle-maven-publish-plugin:${System.getProperty("com.vanniktech.publish.version")}"
+        classpath "com.android.tools.build:gradle:7.1.2"
+        classpath "org.jetbrains.dokka:dokka-gradle-plugin:1.6.10"
     }
 }
