+++ plugin/src/integrationTest/fixtures/passing_android_project/expected/test-artifact-1.0.0.pom
@@ -10,7 +10,7 @@
   <groupId>com.example</groupId>
   <artifactId>test-artifact</artifactId>
   <version>1.0.0</version>
-  <packaging>aar</packaging>
+  <packaging>pom</packaging>
   <name>Gradle Maven Publish Plugin Test Artifact</name>
   <description>Testing the Gradle Maven Publish Plugin</description>
   <url>https://github.com/vanniktech/gradle-maven-publish-plugin/</url>
