+++ plugin/src/integrationTest/fixtures/passing_android_with_kotlin_project/build.gradle
@@ -9,7 +9,7 @@ buildscript {
         classpath "com.vanniktech:gradle-maven-publish-plugin:${System.getProperty("com.vanniktech.publish.version")}"
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.10"
         classpath "org.jetbrains.dokka:dokka-gradle-plugin:1.6.10"
-        classpath "com.android.tools.build:gradle:7.0.0"
+        classpath "com.android.tools.build:gradle:7.1.2"
     }
 }
