+++ plugin/src/integrationTest/kotlin/com/vanniktech/maven/publish/MavenPublishPluginIntegrationTest.kt
@@ -102,29 +102,8 @@ class MavenPublishPluginIntegrationTest {
     val result = executeGradleCommands(TEST_TASK, "--stacktrace")
 
     assertExpectedTasksRanSuccessfully(result)
-    assertExpectedCommonArtifactsGenerated(artifactExtension = "aar")
-    assertPomContentMatches()
-    assertSourceJarContainsFile("com/vanniktech/maven/publish/test/TestActivity.java", "src/main/java")
-  }
-
-  @Test fun generatesArtifactsAndDocumentationOnAndroid711Project() {
-    setupFixture("passing_android_7_1_1_project")
-
-    val result = executeGradleCommands(TEST_TASK, "--stacktrace")
-
-    assertExpectedTasksRanSuccessfully(result)
     assertExpectedCommonArtifactsGenerated("aar", qualifier = "debug")
     assertExpectedCommonArtifactsGenerated("aar", qualifier = "release")
-    assertPomContentMatches()
-  }
-
-  @Test fun generatesArtifactsAndDocumentationOnAndroid711ReleaseVariantProject() {
-    setupFixture("passing_android_7_1_1_release_variant_project")
-
-    val result = executeGradleCommands(TEST_TASK, "--stacktrace")
-
-    assertExpectedTasksRanSuccessfully(result)
-    assertExpectedCommonArtifactsGenerated("aar")
     assertPomContentMatches()
   }
