+++ plugin/src/main/kotlin/com/vanniktech/maven/publish/MavenPublishBasePlugin.kt
@@ -10,7 +10,12 @@ open class MavenPublishBasePlugin : Plugin<Project> {
   override fun apply(project: Project) {
     val gradleVersion = VersionNumber.parse(project.gradle.gradleVersion)
     if (gradleVersion < MIN_GRADLE_VERSION) {
-      error("You need Gradle version 6.6.0 or higher")
+      error("You need Gradle version 7.2.0 or higher")
+    }
+    project.plugins.withId("com.android.library") {
+      if (!project.hasWorkingNewAndroidPublishingApi()) {
+        error("You need AGP version 7.1.2, 7.2.0-beta02, 7.3.0-alpha01 or newer")
+      }
     }
 
     project.rootProject.plugins.apply(MavenPublishRootPlugin::class.java)
@@ -20,7 +25,20 @@ open class MavenPublishBasePlugin : Plugin<Project> {
     project.extensions.create("mavenPublishing", MavenPublishBaseExtension::class.java, project)
   }
 
+  private fun Project.hasWorkingNewAndroidPublishingApi(): Boolean {
+    // All 7.3.0 builds starting from 7.3.0-alpha01 are fine.
+    if (isAtLeastUsingAndroidGradleVersionAlpha(7, 3, 0, 1)) {
+      return true
+    }
+    // 7.2.0 is fine starting with beta 2
+    if (isAtLeastUsingAndroidGradleVersionAlpha(7, 2, 0, 1)) {
+      return isAtLeastUsingAndroidGradleVersionBeta(7, 2, 0, 2)
+    }
+    // Earlier versions are fine starting with 7.1.2
+    return isAtLeastUsingAndroidGradleVersion(7, 1, 2)
+  }
+
   private companion object {
-    val MIN_GRADLE_VERSION = VersionNumber.parse("6.6.0")
+    val MIN_GRADLE_VERSION = VersionNumber.parse("7.2.0")
   }
 }
