+++ plugin/src/main/kotlin/com/vanniktech/maven/publish/MavenPublishPluginExtension.kt
@@ -22,18 +22,6 @@ abstract class MavenPublishPluginExtension(
   var sonatypeHost: SonatypeHost? = defaultSonatypeHost()
 
   /**
-   * The Android library variant that should be published. Projects not using any product flavors, that just want
-   * to publish the release build type can use the default.
-   *
-   * @Since 0.9.0
-   */
-  @Deprecated(
-    "All variants will be published by default. If you need to publish a single variant library " +
-      "use AndroidSingleVariantLibrary on the base plugin"
-  )
-  var androidVariantToPublish: String? = null
-
-  /**
    * Whether release artifacts should be signed before getting published.
    *
    * @Since 0.9.0
