+++ plugin/src/main/kotlin/com/vanniktech/maven/publish/tasks/JavadocJar.kt
@@ -12,30 +12,22 @@ open class JavadocJar : Jar() {
   }
 
   internal companion object {
-    internal fun Project.javadocJarTask(javadocJar: JavadocJarOption, android: Boolean = false): TaskProvider<*>? {
+    internal fun Project.javadocJarTask(javadocJar: JavadocJarOption): TaskProvider<*>? {
       return when (javadocJar) {
         is JavadocJarOption.None -> null
         is JavadocJarOption.Empty -> emptyJavadocJar()
-        is JavadocJarOption.Javadoc -> plainJavadocJar(android)
+        is JavadocJarOption.Javadoc -> plainJavadocJar()
         is JavadocJarOption.Dokka -> dokkaJavadocJar(javadocJar)
       }
     }
 
     private fun Project.emptyJavadocJar(): TaskProvider<*> = tasks.register("emptyJavadocJar", JavadocJar::class.java)
 
-    private fun Project.plainJavadocJar(android: Boolean): TaskProvider<*> {
-      return if (android) {
-        val androidJavadoc = tasks.register("androidJavadoc", AndroidJavadocs::class.java)
-        tasks.register("androidJavadocJar", JavadocJar::class.java) {
-          it.dependsOn(androidJavadoc)
-          it.from(androidJavadoc)
-        }
-      } else {
-        tasks.register("simpleJavadocJar", JavadocJar::class.java) {
-          val task = tasks.named("javadoc")
-          it.dependsOn(task)
-          it.from(task)
-        }
+    private fun Project.plainJavadocJar(): TaskProvider<*> {
+      return tasks.register("simpleJavadocJar", JavadocJar::class.java) {
+        val task = tasks.named("javadoc")
+        it.dependsOn(task)
+        it.from(task)
       }
     }
